Refactor to move the uploading functions out to a module
